--- tests/ai_controls/test_ai_controls_accessibility.py
@@ -1,265 +1,0 @@
-"""
-High Priority Accessibility Tests for AI Controls
-Maps to TestRail cases: C3276002, C3276003, C3276004
-"""
-import logging
-import pytest
-from modules.page_object_prefs import AboutPrefs
-
-
-@pytest.fixture()
-def test_case():
-    return "3276002"
-
-
-def test_ai_features_keyboard_accessible(about_prefs: AboutPrefs):
-    """
-    C3276002 - AI features keyboard accessible
-    Verify that each Option from the AI Settings page is Keyboard accessible
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1-2: Verify elements are keyboard accessible
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Get elements to verify they exist and can be interacted with
-    try:
-        toggle = about_prefs.get_element("ai-controls-toggle")
-        assert toggle is not None, "AI Controls toggle should exist"
-        
-        # Simulate keyboard interaction - in real test, use pynput
-        # For now, we verify the element is accessible
-        logging.info("AI Controls toggle is accessible")
-    except Exception as e:
-        pytest.fail(f"AI Controls toggle should be keyboard accessible: {e}")
-    
-    # Step 3-4: Verify chatbot provider dropdown is keyboard accessible
-    try:
-        provider_dropdown = about_prefs.get_element("ai-control-sidebar-chatbot-select")
-        assert provider_dropdown is not None, "Chatbot provider dropdown should exist"
-        logging.info("Chatbot provider dropdown is accessible")
-    except Exception as e:
-        pytest.fail(f"Chatbot provider dropdown should be keyboard accessible: {e}")
-    
-    # Step 5-6: Verify block all button is keyboard accessible
-    try:
-        block_button = about_prefs.get_element("block-all-ai-enhancements-button")
-        assert block_button is not None, "Block all button should exist"
-        logging.info("Block all AI enhancements button is accessible")
-    except Exception as e:
-        logging.info(f"Block all button may use moz-toggle element: {e}")
-
-
-def test_keyboard_navigation_to_translations_setting(about_prefs: AboutPrefs):
-    """
-    Verify that Translations setting can be reached and modified via keyboard
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        translations = about_prefs.get_element("ai-control-translations-select")
-        assert translations is not None, "Translations dropdown should exist"
-        logging.info("Translations setting is keyboard accessible")
-    except Exception as e:
-        pytest.fail(f"Translations setting should be keyboard accessible: {e}")
-
-
-def test_keyboard_navigation_to_link_preview_setting(about_prefs: AboutPrefs):
-    """
-    Verify that Link Preview setting can be reached and modified via keyboard
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        link_preview = about_prefs.get_element("ai-control-link-preview-select")
-        assert link_preview is not None, "Link Preview dropdown should exist"
-        logging.info("Link Preview setting is keyboard accessible")
-    except Exception as e:
-        logging.info(f"Link Preview setting not available: {e}")
-
-
-def test_keyboard_navigation_to_smart_tab_groups_setting(about_prefs: AboutPrefs):
-    """
-    Verify that Smart Tab Groups setting can be reached and modified via keyboard
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        tab_groups = about_prefs.get_element("ai-control-smart-tab-groups-select")
-        assert tab_groups is not None, "Smart Tab Groups dropdown should exist"
-        logging.info("Smart Tab Groups setting is keyboard accessible")
-    except Exception as e:
-        logging.info(f"Smart Tab Groups setting not available: {e}")
-
-
-
-def test_ai_settings_visible_with_high_contrast(about_prefs: AboutPrefs):
-    """
-    C3276003 - AI settings visible with high contrast
-    Verify that the AI settings page is visible with High Contrast and other themes enabled
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1-3: Verify page elements are visible
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Get elements to verify they can be interacted with
-    try:
-        toggle = about_prefs.get_element("ai-controls-toggle")
-        assert toggle is not None, "Toggle should be visible"
-        logging.info("AI Controls toggle is visible")
-    except Exception as e:
-        pytest.fail(f"AI Controls toggle should be visible: {e}")
-    
-    # Step 4-5: Verify dropdown is visible
-    try:
-        provider_dropdown = about_prefs.get_element("ai-control-sidebar-chatbot-select")
-        assert provider_dropdown is not None, "Dropdown should be visible"
-        logging.info("Chatbot provider dropdown is visible")
-    except Exception as e:
-        pytest.fail(f"Chatbot provider dropdown should be visible: {e}")
-    
-    # Step 6-7: Verify learn more links are visible
-    try:
-        learn_more = about_prefs.get_element("ai-controls-learn-more-link")
-        assert learn_more is not None, "Learn more link should be visible"
-        logging.info("Learn more link is visible")
-    except Exception as e:
-        logging.info(f"Learn more link may not exist in current version: {e}")
-    
-    # Step 8-9: Verify restore button is visible
-    try:
-        restore_button = about_prefs.get_element("ai-controls-restore-button")
-        assert restore_button is not None, "Restore button should be visible"
-        logging.info("Restore button is visible")
-    except Exception as e:
-        logging.info(f"Restore button may not exist: {e}")
-
-
-def test_ai_dropdowns_visible_with_high_contrast(about_prefs: AboutPrefs):
-    """
-    Verify that all AI dropdowns remain visible with high contrast enabled
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        # Verify translations dropdown
-        translations = about_prefs.get_element("ai-control-translations-select")
-        assert translations is not None
-        
-        # Verify chatbot provider dropdown
-        chatbot = about_prefs.get_element("ai-control-sidebar-chatbot-select")
-        assert chatbot is not None
-        
-        logging.info("All AI dropdowns are visible with high contrast")
-    except Exception as e:
-        pytest.fail(f"AI dropdowns should be visible: {e}")
-
-
-def test_ai_notifications_visible_with_high_contrast(about_prefs: AboutPrefs):
-    """
-    Verify that AI feature notifications and cancel buttons are visible with high contrast
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Enable a feature to trigger notification
-    try:
-        about_prefs.set_ai_translations("enabled")
-        logging.info("Feature notification is visible with high contrast")
-    except Exception as e:
-        logging.info(f"Could not verify notification visibility: {e}")
-
-
-def test_ai_controls_read_by_screen_readers(about_prefs: AboutPrefs):
-    """
-    C3276004 - AI controls read by screen readers
-    Verify that the AI settings page is read out loud with Screen readers
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1-2: Verify elements have proper ARIA labels for screen readers
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    try:
-        toggle = about_prefs.get_element("ai-controls-toggle")
-        # In real scenario, we would verify ARIA attributes
-        # e.g., aria-label, aria-described-by, role attributes
-        logging.info("AI Controls toggle has screen reader support")
-    except Exception as e:
-        pytest.fail(f"AI Controls toggle should have screen reader support: {e}")
-    
-    # Step 3-5: Verify notifications have screen reader support
-    try:
-        about_prefs.set_ai_translations("enabled")
-        logging.info("Feature notifications have screen reader support")
-    except Exception as e:
-        logging.info(f"Could not verify notification screen reader support: {e}")
-    
-    # Step 6-7: Verify dropdown options are announced
-    try:
-        provider = about_prefs.get_ai_chatbot_provider()
-        logging.info(f"Chatbot provider option '{provider}' is readable by screen readers")
-    except Exception as e:
-        logging.info(f"Could not verify dropdown screen reader support: {e}")
-    
-    # Step 8-9: Verify restore button is announced
-    try:
-        restore_button = about_prefs.get_element("ai-controls-restore-button")
-        logging.info("Restore button is announced by screen readers")
-    except Exception as e:
-        logging.info(f"Restore button screen reader support: {e}")
-
-
-def test_feature_labels_announced_by_screen_readers(about_prefs: AboutPrefs):
-    """
-    Verify that all feature labels are properly announced by screen readers
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify all major feature labels are accessible
-    features_to_check = [
-        ("ai-control-translations-select", "Translations"),
-        ("ai-control-sidebar-chatbot-select", "AI Chatbot"),
-        ("ai-control-link-preview-select", "Link Previews"),
-        ("ai-control-smart-tab-groups-select", "Smart Tab Groups"),
-        ("ai-control-pdf-alt-text-select", "PDF Alt Text"),
-    ]
-    
-    for element_id, feature_name in features_to_check:
-        try:
-            element = about_prefs.get_element(element_id)
-            # In real scenario, verify ARIA labels
-            logging.info(f"{feature_name} label is accessible to screen readers")
-        except Exception as e:
-            logging.info(f"{feature_name} not available: {e}")
-
-
-def test_killswitch_toggle_announced_by_screen_readers(about_prefs: AboutPrefs):
-    """
-    Verify that the killswitch toggle state is properly announced
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        # Verify toggle has proper aria-label or aria-described-by
-        toggle = about_prefs.get_element("ai-controls-toggle")
-        # Get toggle state and verify it's announced correctly
-        state = about_prefs.get_ai_killswitch_state()
-        logging.info(f"Killswitch state '{state}' is announced by screen readers")
-    except Exception as e:
-        pytest.fail(f"Killswitch toggle should be announced: {e}")
-
-
-def test_dropdown_options_announced_by_screen_readers(about_prefs: AboutPrefs):
-    """
-    Verify that dropdown options are properly announced by screen readers
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        # For each dropdown, verify options are announced
-        # In real scenario, expand dropdown and verify option announcements
-        about_prefs.set_ai_translations("enabled")
-        logging.info("Dropdown options are announced by screen readers")
-    except Exception as e:
-        logging.info(f"Could not verify dropdown option announcement: {e}")
