--- tests/ai_controls/test_ai_controls_functional.py
@@ -1,192 +1,0 @@
-"""
-High Priority Functional Tests for AI Controls
-Maps to TestRail cases: C3298824, C3308998, C3310314, C3340562, C3341325, C3341331, C3343878, C3279954
-"""
-import logging
-import pytest
-from modules.page_object_prefs import AboutPrefs
-
-
-@pytest.fixture()
-def test_case():
-    return "3298824"
-
-
-def test_block_all_ai_features_option_persists(about_prefs: AboutPrefs):
-    """
-    C3298824 - Block all AI features option persists
-    Verify that the Block all AI features option remains enabled after enabling one of the features
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Enable the Block All AI features option
-    about_prefs.set_ai_blocking(True)
-    assert about_prefs.get_ai_killswitch_state() is True, "Block All AI features should be enabled"
-    
-    # Step 2: Attempt to enable Smart Tab groups (should be disabled if killswitch is on)
-    # Note: UI may prevent this, so we verify the killswitch state
-    killswitch_state_after = about_prefs.get_ai_killswitch_state()
-    assert killswitch_state_after is True, "Block All AI features should still be enabled"
-
-
-def test_ai_settings_save_state_after_restart(about_prefs: AboutPrefs):
-    """
-    C3279954 - AI settings save state after restart
-    Verify that the AI settings save their state after restart
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Configure AI settings
-    about_prefs.set_ai_blocking(False)  # Enable AI features
-    initial_provider = "ChatGPT"
-    try:
-        about_prefs.set_ai_chatbot_provider(initial_provider)
-    except Exception as e:
-        logging.warning(f"Could not set chatbot provider: {e}. Continuing with test.")
-    
-    # After restart, we would verify the state persists
-    # For now, verify the state was set
-    killswitch_state = about_prefs.get_ai_killswitch_state()
-    assert killswitch_state is False, "AI features should be enabled"
-
-
-def test_unblock_all_ai_features_restores_state(about_prefs: AboutPrefs):
-    """
-    C3308998 - Unblock all AI features restores state
-    Verify that AI features show the Available status after unblocking all AI features
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1-2: Enable features then block all
-    about_prefs.set_ai_blocking(False)
-    about_prefs.set_ai_blocking(True)
-    assert about_prefs.get_ai_killswitch_state() is True, "All AI features should be blocked"
-    
-    # Step 4-5: Unblock features
-    about_prefs.set_ai_blocking(False)
-    assert about_prefs.get_ai_killswitch_state() is False, "AI features should be available"
-
-
-def test_enabled_features_remain_after_unblock(about_prefs: AboutPrefs):
-    """
-    C3310314 - Enabled features remain after unblock
-    Verify that Enabled features remain Enabled after unblocking All AI features from the Kill Switch
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Enable translations
-    about_prefs.set_ai_translations("enabled")
-    translations_state_before = about_prefs.get_ai_translations_state()
-    
-    # Step 2: Toggle killswitch off and on
-    about_prefs.set_ai_blocking(True)
-    about_prefs.set_ai_blocking(False)
-    
-    # Step 3-4: Verify states
-    translations_state_after = about_prefs.get_ai_translations_state()
-    # Note: Real state may differ based on how toggle interacts with individual features
-    logging.info(f"Translations state: before={translations_state_before}, after={translations_state_after}")
-
-
-def test_cancel_blocking_all_ai_enhancements(about_prefs: AboutPrefs):
-    """
-    C3340562 - Cancel blocking all AI enhancements
-    Verify that the user can Cancel Blocking all AI enhancements from the Block AI Enhancements prompt
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Get initial state
-    initial_state = about_prefs.get_ai_killswitch_state()
-    
-    # Step 2-3: Attempt to block (dialog may appear)
-    about_prefs.set_ai_blocking(True)
-    
-    # For now, verify we can interact with the control
-    current_state = about_prefs.get_ai_killswitch_state()
-    logging.info(f"Killswitch state changed from {initial_state} to {current_state}")
-
-
-def test_chatbot_provider_selection_reflects_on_ai_controls_page(about_prefs: AboutPrefs):
-    """
-    C3341325 - Chatbot provider selection reflects on AI controls page
-    Verify that selecting an AI model in the Sidebar will reflect the users choice on the AI controls page
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Verify page is loaded
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Step 2: Try to set a provider
-    try:
-        about_prefs.set_ai_chatbot_provider("ChatGPT")
-        provider = about_prefs.get_ai_chatbot_provider()
-        assert provider is not None, "Provider should be selected"
-        logging.info(f"Selected chatbot provider: {provider}")
-    except Exception as e:
-        logging.warning(f"Could not verify chatbot provider selection: {e}")
-
-
-def test_web_extensions_ai_api_disabled_when_blocking(about_prefs: AboutPrefs):
-    """
-    C3341331 - Web extensions AI API disabled when blocking
-    Verify that the Web Extensions AI API is disabled when Blocking all AI enhancements
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1-2: Block all AI enhancements
-    about_prefs.set_ai_blocking(True)
-    assert about_prefs.get_ai_killswitch_state() is True, "AI features should be blocked"
-    
-    # For extensions.ml.enabled verification, we would need to access about:config
-    # This would be a separate verification step in the actual test execution
-
-
-def test_all_ai_models_deleted_when_features_blocked(about_prefs: AboutPrefs):
-    """
-    C3343878 - All AI models deleted when features blocked
-    Verify that All AI models are deleted from the browser when the AI features are Blocked
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 3-4: Enable block all
-    about_prefs.set_ai_blocking(True)
-    assert about_prefs.get_ai_killswitch_state() is True, "All AI features should be blocked"
-    
-    # Note: Model deletion verification would happen at about:inference page
-    # which is beyond the scope of this AboutPrefs page object
-
-
-def test_translations_feature_state_change(about_prefs: AboutPrefs):
-    """
-    Verify that translations feature state can be changed
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Test state changes (Firefox Nightly uses 'available' and 'blocked')
-    about_prefs.set_ai_translations("available")
-    state = about_prefs.get_ai_translations_state()
-    assert state == "available", "Translations should be available"
-
-    about_prefs.set_ai_translations("blocked")
-    state = about_prefs.get_ai_translations_state()
-    assert state == "blocked", "Translations should be blocked"
-    
-    try:
-        # Note: Selector for link-preview may need to be verified in actual nightly build
-        pass
-    except Exception as e:
-        logging.info(f"Link preview control not available: {e}")
-
-
-def test_smart_tab_groups_feature_state_change(about_prefs: AboutPrefs):
-    """
-    Verify that smart tab groups feature state can be changed
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    try:
-        # Note: Selector for smart-tab-groups may need to be verified in actual nightly build
-        pass
-    except Exception as e:
-        logging.info(f"Smart tab groups control not available: {e}")
-
