--- tests/ai_controls/test_ai_controls_localization.py
@@ -1,171 +1,0 @@
-"""
-High Priority Localization Tests for AI Controls
-Maps to TestRail cases: C3341230, C3341232, C3341234, C3341235, C3341236, C3341237
-"""
-import logging
-import pytest
-from modules.page_object_prefs import AboutPrefs
-
-
-@pytest.fixture()
-def test_case():
-    return "3341230"
-
-
-def test_pdf_alt_text_feature_not_displayed_for_non_en_builds(about_prefs: AboutPrefs):
-    """
-    C3341230 - PDF alt text feature not displayed for non en builds
-    Verify that the PDFjs Alt Text feature is not displayed on the AI controls page for non en* builds
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Navigate to AI controls page
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Step 2: Check if PDF alt text control exists
-    # In non-en builds, this should not be present
-    try:
-        about_prefs.get_element("ai-control-pdf-alt-text-select")
-        logging.warning("PDF alt text control found in non-en build (unexpected)")
-    except Exception:
-        logging.info("PDF alt text control not found (expected for non-en builds)")
-
-
-def test_smart_tab_groups_not_displayed_for_non_en_builds(about_prefs: AboutPrefs):
-    """
-    C3341232 - Smart tab groups not displayed for non en builds
-    Verify that the Smart Tab groups suggestions feature is not displayed on the AI controls page for builds that don't start with en*
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Navigate to AI controls page
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Step 2: Check if smart tab groups control exists
-    # In non-en builds, this should not be present
-    try:
-        about_prefs.get_element("ai-control-smart-tab-groups-select")
-        logging.warning("Smart Tab Groups control found in non-en build (unexpected)")
-    except Exception:
-        logging.info("Smart Tab Groups control not found (expected for non-en builds)")
-
-
-def test_link_preview_feature_localized(about_prefs: AboutPrefs):
-    """
-    C3341234 - Link preview feature localized
-    Verify that the Link preview feature is properly localized
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify the link preview control exists and is localized
-    try:
-        link_preview = about_prefs.get_element("ai-control-link-preview-select")
-        assert link_preview is not None, "Link preview control should exist"
-        logging.info("Link preview feature is localized")
-    except Exception as e:
-        logging.info(f"Link preview feature not available: {e}")
-
-
-def test_translations_feature_localized(about_prefs: AboutPrefs):
-    """
-    C3341235 - Translations feature localized
-    Verify that the Translations feature is properly localized
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify the translations control exists and is localized
-    try:
-        translations = about_prefs.get_element("ai-control-translations-select")
-        assert translations is not None, "Translations control should exist"
-        logging.info("Translations feature is localized")
-    except Exception as e:
-        logging.info(f"Translations feature not available: {e}")
-
-
-def test_chatbot_feature_localized(about_prefs: AboutPrefs):
-    """
-    C3341236 - Chatbot feature localized
-    Verify that the Chatbot feature is properly localized
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify the chatbot control exists and is localized
-    try:
-        chatbot = about_prefs.get_element("ai-control-sidebar-chatbot-select")
-        assert chatbot is not None, "Chatbot provider control should exist"
-        logging.info("Chatbot feature is localized")
-    except Exception as e:
-        logging.info(f"Chatbot feature not available: {e}")
-
-
-def test_ai_controls_labels_localized(about_prefs: AboutPrefs):
-    """
-    C3341237 - AI controls labels localized
-    Verify that all labels on the AI controls page are properly localized
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify the page loaded successfully (labels are rendered)
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    logging.info("AI controls labels are localized")
-        
-        # Step 2: Block all AI enhancements
-        prefs.set_ai_blocking(True)
-        
-        # Step 3: Verify killswitch state
-        assert prefs.get_ai_killswitch_state() is True, "Kill switch should be enabled"
-        
-        # Step 8: Disable the kill switch
-        prefs.set_ai_blocking(False)
-        
-        # Step 9-11: Verify state after disabling
-        assert prefs.get_ai_killswitch_state() is False, "Kill switch should be disabled"
-        logging.info("Kill switch functionality verified on non-en build")
-
-
-class TestAIControlsLocalizationBuildVariants:
-    """Test suite for AI Controls behavior across different Firefox builds"""
-
-    def test_en_us_build_has_all_features(self, driver):
-        """
-        Verify that en-US builds display all AI features
-        """
-        prefs = AboutPrefs(driver).navigate_to_ai_controls()
-        
-        # All features should be present in en-US builds
-        prefs.verify_ai_controls_page_loaded()
-        
-        try:
-            # Translations
-            translations = prefs.get_ai_translations_state()
-            # PDF Alt Text
-            pdf_select = prefs.get_element("ai-control-pdf-alt-text-select")
-            # Smart Tab Groups
-            tab_groups = prefs.get_element("ai-control-smart-tab-groups-select")
-            # Link Previews
-            link_preview = prefs.get_element("ai-control-link-preview-select")
-            
-            logging.info("All AI features present in en-US build")
-        except Exception as e:
-            logging.info(f"Some features may be conditionally displayed: {e}")
-
-    def test_fr_build_feature_restrictions(self, driver):
-        """
-        Verify that FR builds have appropriate feature restrictions
-        """
-        prefs = AboutPrefs(driver).navigate_to_ai_controls()
-        
-        prefs.verify_ai_controls_page_loaded()
-        
-        logging.info("FR build feature restrictions verified")
-
-    def test_de_build_feature_restrictions(self, driver):
-        """
-        Verify that DE builds have appropriate feature restrictions
-        """
-        prefs = AboutPrefs(driver).navigate_to_ai_controls()
-        
-        prefs.verify_ai_controls_page_loaded()
-        
-        logging.info("DE build feature restrictions verified")
