--- tests/ai_controls/test_ai_controls_policies.py
@@ -1,137 +1,0 @@
-"""
-High Priority Enterprise Policies Tests for AI Controls
-Maps to TestRail cases: C3277712, C3277713, C3277714, C3277716, C3279955, C3298825
-"""
-import logging
-import pytest
-from modules.page_object_prefs import AboutPrefs
-
-
-@pytest.fixture()
-def test_case():
-    return "3277712"
-
-
-def test_disable_all_ai_features_via_policies(about_prefs: AboutPrefs):
-    """
-    C3277712 - Disable all AI features via policies
-    Verify that the user can Disable all AI feature settings via policies
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Step 1: Navigate to AI controls
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Steps 2-3: When policies disable all features, the controls should be in a disabled state
-    # This test validates the presence of the controls even when policy-enforced
-    try:
-        toggle_state = about_prefs.get_ai_killswitch_state()
-        logging.info(f"Killswitch state under policy: {toggle_state}")
-    except Exception as e:
-        logging.info(f"Expected behavior - controls may be hidden/disabled under strict policy: {e}")
-
-
-def test_disable_only_chatbot_via_policies(about_prefs: AboutPrefs):
-    """
-    C3277713 - Disable only chatbot via policies
-    Verify that the user can disable only the Chatbot feature via policies
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify page loads and other controls are accessible
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Try to access translations (should work if only chatbot is disabled)
-    try:
-        state = about_prefs.get_ai_translations_state()
-        logging.info(f"Translations feature state: {state}")
-    except Exception as e:
-        logging.info(f"Translations control status: {e}")
-
-
-def test_disable_only_smart_tab_groups_via_policies(about_prefs: AboutPrefs):
-    """
-    C3277714 - Disable only smart tab groups via policies
-    Verify that the user can disable only the Smart Tab groups feature via policies
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify page loads
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    logging.info("Smart Tab Groups policy test - verifying other features remain accessible")
-
-
-def test_disable_only_link_previews_via_policies(about_prefs: AboutPrefs):
-    """
-    C3277716 - Disable only link previews via policies
-    Verify that the user can disable only the Link previews feature via policies
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify page loads
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    logging.info("Link Previews policy test - verifying other features remain accessible")
-
-
-def test_enable_ai_features_when_policy_changed(about_prefs: AboutPrefs):
-    """
-    C3279955 - Enable AI features when policy changed
-    Verify that the AI features can be enabled if the policy is changed or removed
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # After policy change, features should be enabled
-    # Verify the controls are accessible and enabled
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    killswitch_state = about_prefs.get_ai_killswitch_state()
-    logging.info(f"Killswitch state after policy change: {killswitch_state}")
-
-
-def test_disable_only_translations_via_policies(about_prefs: AboutPrefs):
-    """
-    C3298825 - Disable only translations via policies
-    Verify that the user can disable only the Translations feature via policies
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Verify page loads
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    # Try to access translations - should show removed/disabled state
-    try:
-        state = about_prefs.get_ai_translations_state()
-        logging.info(f"Translations state with policy: {state}")
-    except Exception as e:
-        logging.info(f"Translations unavailable (expected under policy): {e}")
-
-
-def test_policy_locked_settings_not_modifiable(about_prefs: AboutPrefs):
-    """
-    Verify that settings locked by policy cannot be modified by user
-    
-    When GenerativeAI policy has Locked: true, users should not be able
-    to toggle or change any settings
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # Attempt to change settings when policy is locked
-    # UI should prevent modification or show disabled state
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    logging.info("Policy locking test - verifying controls are in expected states")
-
-
-def test_partial_policy_enforcement(about_prefs: AboutPrefs):
-    """
-    Verify that partial policies (some features disabled, some enabled)
-    work correctly together
-    """
-    about_prefs.navigate_to_ai_controls()
-    
-    # With mixed policies, some controls should be available, others disabled
-    about_prefs.verify_ai_controls_page_loaded()
-    
-    logging.info("Partial policy test - verifying selective feature enforcement")
